rename to diskhashmap
