+++ .github/workflows/release.yml
@@ -33,8 +33,8 @@ jobs:
       - name: Install Rust
         uses: dtolnay/rust-toolchain@stable
 
-      - name: Publish opendiskmap
-        run: cargo publish -p opendiskmap --token ${{ secrets.CARGO_REGISTRY_TOKEN }}
+      - name: Publish diskhashmap
+        run: cargo publish -p diskhashmap --token ${{ secrets.CARGO_REGISTRY_TOKEN }}
 
       - name: Publish diskdashmap
         run: cargo publish -p diskdashmap --token ${{ secrets.CARGO_REGISTRY_TOKEN }}
