+++ diskhashmap/benches/u64_key_benchmark.rs
@@ -1,6 +1,6 @@
 use criterion::{Criterion, black_box, criterion_group, criterion_main};
-use opendiskmap::disk_map::DiskHashMap;
-use opendiskmap::types::{Bytes, Native};
+use diskhashmap::disk_map::DiskHashMap;
+use diskhashmap::types::{Bytes, Native};
 use rand::{Rng, RngCore};
 use rustc_hash::FxBuildHasher;
 use std::time::Duration;
@@ -8,7 +8,7 @@ use tempfile::tempdir;
 
 // Type alias for our Mmap-backed HashMap with u64 keys and Vec<u8> values
 type HashMapMmapU64 =
-    DiskHashMap<Native<u64>, Bytes, opendiskmap::byte_store::MMapFile, FxBuildHasher>;
+    DiskHashMap<Native<u64>, Bytes, diskhashmap::byte_store::MMapFile, FxBuildHasher>;
 
 /// Generates a vector of key-value pairs for benchmarking.
 /// Keys are random u64 values, and values are random byte vectors.
