+++ diskhashmap/examples/native_types_demo.rs
@@ -1,5 +1,5 @@
-use opendiskmap::byte_store::MMapFile;
-use opendiskmap::{DiskHashMap, Native, Result, Str};
+use diskhashmap::byte_store::MMapFile;
+use diskhashmap::{DiskHashMap, Native, Result, Str};
 use rustc_hash::FxBuildHasher;
 use tempfile::tempdir;
