+++ diskhashmap/src/buffers.rs
