+++ diskhashmap/src/byte_store.rs
