+++ diskhashmap/src/disk_map.rs
