+++ diskhashmap/src/entry.rs
