+++ diskhashmap/src/error.rs
