+++ diskhashmap/src/fixed_buffers.rs
