+++ diskhashmap/src/heap.rs
