+++ diskhashmap/src/lib.rs
