+++ diskhashmap/src/storage.rs
