+++ diskhashmap/src/types.rs
